Factor out threading abstraction for BatchWorkers and use normal Thread subclass on Apple platforms
